+++ spek-runtime/src/commonMain/kotlin/org/spekframework/spek2/runtime/util/Base64.kt
@@ -1,4 +1,4 @@
-package org.spekframework.spek2.runtime.scope
+package org.spekframework.spek2.runtime.util
 
 expect object Base64 {
     fun encodeToString(text: String): String
